% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{avoidZero}
\alias{avoidZero}
\title{Get rid of zeros in a numeric vector}
\usage{
avoidZero(x, fac = 10)
}
\arguments{
\item{x}{numeric values as input}

\item{fac}{numeric a factor of scale in order to get a "small" value to 
replace zeros}
}
\value{
a vector of value with zeros replaced.
}
\description{
Get rid of zeros in a numeric vector before taking 
the logarithm of them. We basically replace the "zeros" with a negligible small 
 value in order to avoid NaN upon the log-transformation.
}
