% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISAplate.R
\docType{class}
\name{elisa_batch-class}
\alias{elisa_batch-class}
\title{S4 class definition of an elisa_batch object}
\arguments{
\item{batchID}{characters to specify the batch}

\item{runs}{list of elisa_run objects}

\item{num.runs}{numeric the number of elisa_runs in this batch}

\item{pars}{numeric the actually parameters for the fitting.
for example for the 5pl they are c(a, d,xmid, scal, g).}

\item{model.fit}{list intend to contain information for the 
fitting of nls.lm. But not using it now.}

\item{model.name}{characters of either the 5pl (5-parameter) or 4pl 
(4-parameter) logistic function}

\item{range.ODs}{numeric the min and max ODs}

\item{normFactor}{numeric the batch normalization factor ("S").}
}
\description{
\code{\link{elisa_batch-class}} define the S4 class of an elisa_batch object
}
\details{
defining the S4 class of the elisa_batch.
This holds the data for elisa batch. 
	It contains one or many elisa_run objects.
}
\section{Slots}{

\describe{
\item{\code{batchID}}{character}

\item{\code{desc}}{character}

\item{\code{runs}}{list}

\item{\code{num.runs}}{numeric}

\item{\code{pars}}{numeric}

\item{\code{model.fit}}{list}

\item{\code{model.name}}{character}

\item{\code{range.ODs}}{numeic}

\item{\code{normFactor}}{numeric}
}}

\examples{
elisa_batch();
}
\seealso{
\code{\link[minpack.lm]{nls.lm}} \code{\link{elisa_plate-class}} \code{\link{elisa_run-class}} 
\code{\link{elisa_batch-class}}
}
\concept{ELISA}
