% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISAplate.R
\docType{class}
\name{elisa_run-class}
\alias{elisa_run-class}
\title{S4 class definition of an elisa_run object}
\arguments{
\item{batchID}{characters to specify the batch}

\item{desc}{characters for the data/experiment information}

\item{plates}{list of elisa_plates}

\item{num.plates}{numeric the number of plates in this run}

\item{date}{characters for the date of running the ELISA measurements}
}
\description{
\code{\link{elisa_run-class}} defines the S4 class of elisa_run
}
\details{
defining the S4 class of the elisa_run object.
This is list to hold the data for each elisa run. 
	It contains one or many elisa plate objects.
}
\section{Slots}{

\describe{
\item{\code{batchID}}{character}

\item{\code{plates}}{list}

\item{\code{desc}}{character}

\item{\code{num.plates}}{numeric}

\item{\code{date}}{character}

\item{\code{range.ODs}}{numeric}
}}

\examples{
elisa_run();
}
\seealso{
\code{\link{nls.lm}}
}
\concept{ELISA}
