% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{f5pl}
\alias{f5pl}
\title{The five-parameter logistic function}
\usage{
f5pl(pars, x)
}
\arguments{
\item{pars}{numeric the parameters of the 5pl (or 4pl). It has the following content: [a, d, xmid, scal, g].}

\item{x}{numeric the log-transformed x value(s).}
}
\value{
the 5pl function value(s).
}
\description{
read in the paramters and the independent variable value(s), and then return the 
	5pl function value(s). For the 4pl model, set g to be 1.
}
\details{
The function has the following form \cr
    f(x)=a+(d-a)/((1+exp((xmid-x)/scal))^g)
}
