% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{inv.f5pl}
\alias{inv.f5pl}
\title{The inverse of the 5-parameter logistic function}
\usage{
inv.f5pl(pars, y)
}
\arguments{
\item{pars}{the parameters of the function. It has the following content: [a, d, xmid, scale, g].}

\item{y}{the value to be reverse calculated.}
}
\value{
the value of the reverse function
}
\description{
The inverse function of the 5pl. Set the value of g to be 1, if the 4pl
is of interest.
}
