% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISAplate.R
\docType{methods}
\name{load.ODs}
\alias{load.ODs}
\alias{load.ODs,elisa_plate-method}
\title{Function to load OD data}
\usage{
load.ODs(x, plate.header, plate.data, plate.blank, annotation, ....)

\S4method{load.ODs}{elisa_plate}(x, plate.header, plate.data, plate.blank,
  annotation)
}
\arguments{
\item{x}{the elisa_plate object to load data into}

\item{plate.header}{characters}

\item{plate.data}{data.frame OD readings}

\item{plate.blank}{data.frame OD blank readings}

\item{annotation}{data.frame annotation to guide reading.}

\item{....}{other parameters that will help reading data.}
}
\description{
Generic function to load OD data into an elisa_plate 
	object
}
\details{
It loads OD data into an elisa_plate object. The data
usually read int from design file, annotation file, OD file and 
standard concentration data.
}
\section{Methods (by class)}{
\itemize{
\item \code{elisa_plate}: to load ODs to an elisa_plate object
}}

