% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{prepareInitsLM}
\alias{prepareInitsLM}
\title{Prepare initial values for fitting shifts}
\usage{
prepareInitsLM(batches, ref.batch = 1)
}
\arguments{
\item{batches}{list of elisa_batch data}

\item{ref.batch}{numeric the index of the reference batch. It is 1
by default.}
}
\value{
a data list contain the following elements,
\itemize{
\item {inits, the initial values for the standard curves of all the plates\cr}
	
\item {ref.ibatch, the index of the reference batch\cr}
	{
		This one is specified by the input ref.batch.
	}
\item {ref.irun,the index of the reference run\cr}

\item {ref.index,the index of the reference line in the order
		of the inits vector\cr}

}
}
\description{
Generate the initial values for fitting shifts with 
a model of the 5-parameter logistic function.
}
\details{
This is a more complicated way to prepare the initials for shifting.
}
