% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{prepareRegInput}
\alias{prepareRegInput}
\title{Prepare the input for regressoin}
\usage{
prepareRegInput(batches)
}
\arguments{
\item{batches}{list of the ELISA data arranged in 
batches. Each element of list contains a batch (list)
data, and each batch contains one or many the elisa_run 
objects \code{\link{elisa_plate}}}
}
\value{
list of data that will feed in to do regression.
}
\description{
Prepare the input data to feed in the fitting.
}
\seealso{
\code{\link{elisa_plate}} \code{\link{elisa_batch}}\code{\link{elisa_run}}
}
