% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{rangeOD}
\alias{rangeOD}
\title{Get the OD ranges (min/max)}
\usage{
rangeOD(batches)
}
\arguments{
\item{batches}{list of batches data}
}
\description{
Going through the list of  batches to get the OD range (min and max)
}
