% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISAtools_IO.R
\name{read.plate}
\alias{read.plate}
\title{Read the single ELISA OD plate}
\usage{
read.plate(ODs, annotation, batchID, expID)
}
\arguments{
\item{ODs}{characters containing data of ODs for one plate}

\item{annotation}{list of data containing annotations of the plate}

\item{batchID}{characters specifying the batchID read from the design file}

\item{expID}{characters specifying the expID or plateID read from the design file}
}
\value{
an object of elisa_plate holding data and annotations for 
	a single plate.
}
\description{
Read the individual ELISA plate to parse the ODs.
}
\details{
The input is a text file imported from the sdf file.  
We only read the first section with both the OD and blank file.
The OD data are read in according to the annotation file.
}
