% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISAtools.R
\docType{package}
\name{ELISAtools-package}
\alias{ELISAtools}
\alias{ELISAtools-package}
\title{ELISA data analysis with batch correction}
\description{
An R package to run ELISA data analysis 
	with the ability to do batch correction/normalization
}
\details{
This package is developed to run analysis of ELISA data. 
	First, the calibration data are used to fit either the five- or
	four-parameter logistic model. Then the fitted model is
	used to predict the concentrations of unknown samples.
	If the batches of calibration data exist, 
	the correction/normalization could be done. The corrected
	calibration curve are then used for predication.

		Please refer to the vignettes to see details.
}
\references{
Feng, et al 2018 \url{https://doi.org/10.1101/483800}
}
\author{
\strong{Maintainer}: Feng Feng \email{ffeng@BU.edu}

}
